requirements.txt
mysqlclient==2.2.4
Django==4.2.2
django-cors-headers==4.1.0
django-crontab==0.7.1
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0